fixed button being created even is a role doesnt have necessary privileges

+++ src/admin_dashboard.py
@@ -8,7 +8,7 @@ from db import db_manager
 from generic_dialog_builders import build_generic_add_dialog, build_generic_delete_dialog, build_generic_get_dialog, \
     build_generic_update_dialog
 from src.update_dialog_builder import update_dialog_builders
-from ui_common import show_all, count_rows, custom_query, get_user_tables, get_all_views
+from ui_common import show_all, count_rows, custom_query, get_user_tables, get_all_views, check_user_privileges
 
 
 def build_dashboard(user, on_logout):
@@ -43,20 +43,26 @@ def build_dashboard(user, on_logout):
                     for lbl in labels:
                         with ui.tab_panel(lbl):
                             with ui.row().classes('full-width q-gutter-sm'):
+                                # Check user privileges
+                                privileges = check_user_privileges(db_manager.conn, lbl)
                                 # Buttons go HEREEEEEE
                                 # Add button with dialogggg
-                                add_builder = add_dialog_builders.get(lbl, build_generic_add_dialog)
-                                add_dlg = add_builder(db_manager.conn, lbl)
-                                ui.button(f'Add to {lbl}', on_click=add_dlg.open)
-                                del_builder = delete_dialog_builders.get(lbl, build_generic_delete_dialog)
-                                del_dlg = del_builder(db_manager.conn, lbl)
-                                ui.button(f'Delete from {lbl}', on_click=del_dlg.open)
+                                if privileges.get('INSERT', False):
+                                    add_builder = add_dialog_builders.get(lbl, build_generic_add_dialog)
+                                    add_dlg = add_builder(db_manager.conn, lbl)
+                                    ui.button(f'Add to {lbl}', on_click=add_dlg.open)
+                                if privileges.get('DELETE', False):
+                                    del_builder = delete_dialog_builders.get(lbl, build_generic_delete_dialog)
+                                    del_dlg = del_builder(db_manager.conn, lbl)
+                                    ui.button(f'Delete from {lbl}', on_click=del_dlg.open)
+
                                 get_builder = get_dialog_builders.get(lbl, build_generic_get_dialog)
                                 get_dlg = get_builder(db_manager.conn, lbl, result_areas)
                                 ui.button(f'Filter from {lbl}', on_click=get_dlg.open)
-                                upd_builder = update_dialog_builders.get(lbl, build_generic_update_dialog)
-                                upd_dlg = upd_builder(db_manager.conn, lbl)
-                                ui.button(f'Update {lbl}', on_click=upd_dlg.open)
+                                if privileges.get('UPDATE', False):
+                                    upd_builder = update_dialog_builders.get(lbl, build_generic_update_dialog)
+                                    upd_dlg = upd_builder(db_manager.conn, lbl)
+                                    ui.button(f'Update {lbl}', on_click=upd_dlg.open)
 
                             with ui.card().classes('full-width'):
                                 ui.label('Выполнить произвольный запрос:').classes('text-weight-bold')
